requirements.txt
Django==2.1.4
djangorestframework==3.9.0
djangorestframework-csv==2.1.0
django-filter==2.0.0
Markdown==3.0.1
pytest-django==3.4.4
psycopg2==2.7.6.1
gunicorn==19.9.0
